Add expose! method to pre-fetch resources
This implementation is modeled from RSpec's let and let! methods,
wherein the bang variation simply adds a before filter to pre-fetch and
cache the block.

The expose! method does take an options hash, which is passed on the
before_filter itself. This allows for only some actions to do the
caching, i.e. with the :only option.