Don't reevaluate a nil exposure block
When the exposure block returns a nil, that value should be returned
instead of re-running the exposure block when the method is re-invoked.
If the exposure encapsulates a long-running operation, you don't want
that to run multiple times.

expose(:foo) { sleep 10; nil }
Bechmark.realtime { foo; foo} = 20 secs, instead of 10 seconds