Merge jgdavey/nunu
Unfortunately, it appears we diverged enough to necessitate a squashed
merge of jgdavey/nunu.

Another side-effect of this merge is the removal of the recently added
DecentExposure::Configuration class. Josh was approaching configuration
as well, but in a different manner. While I still want to provide
top-level configuration, rather than just exposure-specific
configuration, Josh's work was more extensive and our use of nouns in
the namespace conflicted, so I've opted to use his work as the basis for
mine.