Fix nested exposures under Rails 3.1
Fixes the problem demonstrated in:

  https://github.com/acumenbrands/decent_exposure_bug

and alluded to in some comments on #67.

When nested exposures mapping to a has_many association are
instantiated, the fast-path code introduced in 6e913c3d causes the
parent association to return an Array instead of an
ActiveRecord::Relation. This caused operations on the child to fail with
"undefined method `new' for #<Array:...>".

This commit adjusts the fast-path code to only return the scope if the
scope is already a Relation.