Shallow routes with lazy eval'ed ancestors
If there is a shallow route/exposure, this overrides the ancestor
exposure on member routes with the belongs_to association with
the same name as the ancestor route.

This, however, does not work long term since the _exposures are
on the controller class. An override here make a permanent change
to something that lives on the class. Enabling turbolinks in
development demonstrates this.