Eager loads shallow ancestors
Since overridding the exposure itself isn't an option, I
decided just to eager load the shallow ancestor as a proof
of concept.

Still need to write some tests to demonstate this behavior,
but it works in my demo project by adding `shallow: true`
to the exposure that has the shallow routes.

    # routes
    resources :doctor do
      resources :patients, shallow: true
    end

    # PatientsController
    expose(:doctor)
    expose(:patients, ancestor: :doctor, shallow: true)
    expose(:patient)