assign_attributes? on Strong Parameters should use params
The standard on strong_parameters is to use the follow syntax to check
the params:

expose(:post, attributes: :post_params)

def post_params
  params.require(:post)
end

However if we touch the attributes without having a full post on
parameters we would get a `ActionController::ParameterMissing`. In order
to fix that method assign_attributes? on strong parameter strategy
should check against the params instead of attributes

[fixes #99]